split blocks into separate module
